requirements.txt
langchain
openai
pypdf
python_dotenv
tiktoken
